+++ doc/changelog.rst
@@ -4,7 +4,8 @@ Changelog
 Development Version
 -------------------
 
-- 
+- ENH add iteration rate and estimated time to completion to progress bar
+- ENH change ETA format to hh:mm:ss for better readability
 
 Version 0.4.9
 -------------